switch to mars_time package
