+++ mcstools/__init__.py
@@ -1,7 +1,6 @@
 from mcstools import preprocess
 from mcstools.data_path_handler import FilenameBuilder
 from mcstools.loader import L1BLoader, L2Loader
-from mcstools.util.mars_time import MarsDate
 
 __all__ = [
     "FilenameBuilder",
