+++ setup.cfg
@@ -23,9 +23,8 @@ python_requires = >=3.8
 install_requires =
     pandas == 1.4.3
     dask == 2022.6.1
-    marstime == 0.5.6
     numpy==1.23.0
-    scipy==1.8.1
+    scipy==1.11.4
     coloredlogs==15.0.1
     cmcrameri==1.5
     click==8.1.3
